Added New Tab to Keyworddetail: Authors, which reuses the recently implemented authors tab from case studies as a new component, resolves #138
